requirements.txt
asgiref==3.8.1
bleach==3.3.1
Django==5.0.4
django-appconf==1.0.6
django-environ==0.11.2
django-imagekit==5.0.0
django-summernote==0.8.20.0
gunicorn==22.0.0
packaging==24.1
pilkit==3.0
pillow==10.3.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
webencodings==0.5.1
wheel==0.43.0
